[ADD] monorepo command group
