+++ tigr81/tigr81/commands/scaffold/scaffold.py
@@ -4,8 +4,6 @@ from cookiecutter.main import cookiecutter
 import typer
 
 from tigr81 import LOCAL_REPO_LOCATION, REPO_LOCATION
-from tigr81.commands.scaffold import manifest
-from tigr81.commands.scaffold.manifest import Manifest
 from tigr81.commands.scaffold.project_template import (
     Dependency,
     ProjectTemplate,
@@ -15,20 +13,7 @@ from tigr81.commands.scaffold.project_template import (
 
 import pathlib as pl
 
-from tigr81.utils.read_yaml import read_yaml
 
-
-app = typer.Typer()
-
-
-@app.callback()
-def callback():
-    """
-    Scaffold project templates
-    """
-
-
-@app.command()
 def scaffold(
     project_type: ProjectTypeEnum,
     default: bool = typer.Option(
@@ -69,36 +54,6 @@ def scaffold(
     )
 
 
-MANIFEST_FILE_NAME = "manifest.yml"
-
-@app.command("monorepo")
-def scaffold_monorepo():
-    manifest_dct = read_yaml(MANIFEST_FILE_NAME)  
-
-    manifest = Manifest(**manifest_dct)
-
-    PROJECT_TEMPLATE_LOCATION = REPO_LOCATION
-    checkout = "develop"
-
-    if os.getenv("TIGR81_ENVIRONMENT") == "local":
-        PROJECT_TEMPLATE_LOCATION = LOCAL_REPO_LOCATION.as_posix()
-        checkout = None
-
-    typer.echo(
-        f"Scaffolding a monorepo project from {pl.Path.cwd() / MANIFEST_FILE_NAME}"
-    )
-
-    for component in manifest.components:
-        typer.echo(f"Scaffolding {manifest.relative_path / component.relative_path / component.project_options.name} component...")
-        cookiecutter(
-            template=PROJECT_TEMPLATE_LOCATION,
-            output_dir=manifest.relative_path / component.relative_path,
-            no_input=True,
-            extra_context=component.extra_content,
-            checkout=checkout,
-            directory=f"project_templates/{component.project_type}",
-        )   
-
 if __name__ == "__main__":
     project_type = "fastapi"
     PROJECT_TEMPLATE_LOCATION = REPO_LOCATION
@@ -124,12 +79,6 @@ if __name__ == "__main__":
             Dependency(name="fresko", relative_path="../../fresko"),
         ]
     )
-
-    # extra_content = project_template.project_options.momdel_dump()
-    # extra_content["dependencies"] = {
-    #     {"name": "pippo", "version": "1.1.1"},
-    #     {"name": "fresko", "version": "2.2.2"}
-    # }
 
     print(project_template.extra_content)
